190 more notificaitons, check mph before drive state transition, pedals processing task, malloc and stack overflow hooks
