+++ Core/Inc/queues.h
@@ -22,8 +22,8 @@ typedef struct {
 extern osMessageQueueId_t imu_queue;
 
 typedef struct {
-	uint16_t brake_value; /* 0-1 */
-	uint16_t accelerator_value; /* 0-1 */
+	uint16_t brake_value;
+	uint16_t accelerator_value; /* 0-100 */
 } pedals_t;
 
 extern osMessageQueueId_t pedal_data_queue;
