+++ Makefile
@@ -1,5 +1,5 @@
 ##########################################################################################################################
-# File automatically-generated by tool: [projectgenerator] version: [4.3.0-B58] date: [Wed Jun 05 04:33:49 GMT 2024] 
+# File automatically-generated by tool: [projectgenerator] version: [4.2.0-B44] date: [Sun Aug 04 16:29:12 EDT 2024] 
 ##########################################################################################################################
 
 # ------------------------------------------------
@@ -54,7 +54,6 @@ Core/Src/dti.c \
 Core/Src/state_machine.c \
 Core/Src/bms.c \
 Core/Src/nero.c \
-Core/Src/torque.c \
 Core/Src/pdu.c \
 Core/Src/mpu.c \
 Core/Src/steeringio.c \
