+++ cerberus.ioc
@@ -66,8 +66,10 @@ Dma.ADC3.0.RequestParameters=Instance,Direction,PeriphInc,MemInc,PeriphDataAlign
 Dma.Request0=ADC3
 Dma.Request1=ADC1
 Dma.RequestsNb=2
-FREERTOS.IPParameters=Tasks01,configUSE_PREEMPTION
+FREERTOS.IPParameters=Tasks01,configUSE_PREEMPTION,configUSE_MALLOC_FAILED_HOOK,configCHECK_FOR_STACK_OVERFLOW
 FREERTOS.Tasks01=defaultTask,24,128,StartDefaultTask,Default,NULL,Dynamic,NULL,NULL
+FREERTOS.configCHECK_FOR_STACK_OVERFLOW=1
+FREERTOS.configUSE_MALLOC_FAILED_HOOK=1
 FREERTOS.configUSE_PREEMPTION=0
 File.Version=6
 GPIO.groupedBy=Group By Peripherals
